consolidate reducers
